feat: modify datails page, components
