+++ public/movie-details.html
@@ -15,6 +15,6 @@
             <details-textbox></details-textbox>
         </div>
     </div>
-    <script type="module" src="../src/components/details/details.js"></script>
+    <script type="module" src="../src/pages/movie-details.js"></script>
 </body>
 </html>
