+++ src/pages/movie-details.js
@@ -1,49 +1,22 @@
+import '../components/details/details-preview.js'
+import '../components/details/details-textbox.js'
 
+const queryString = window.location.search;
+const params = new URLSearchParams(queryString);
 
-// 상세로 넘오는 클릭 이벤트 발생시 호툴하는 함수에 데이터
-// 페이지가 로드되면 실행되니까
-// api를 호출해서 받아온다.
-// 밸류를 받아서 파라미터로 실행
-// 호출 값으로 넣어서 api데이터 당겨오기 
+// const imdb_id = params.get('imdb_id'); //페이지연결할때 살리기
+const imdb_id = 'tt2294629'// imdb_id_dummy
 
-
-// ((리터럴로)) -> 파라미터 
-// http://www.omdbapi.com/?apikey=33c97183&t=Aladdin
-// http://www.omdbapi.com/?apikey=33c97183&t=frozen
-// http://www.omdbapi.com/?apikey=33c97183&y=2019
-// http://www.omdbapi.com/?apikey=33c97183&type=tt0103639
-// http://www.omdbapi.com/?apikey=33c97183&t=Aladdin&y=2019
-async function getApiData(){
-    const API_URL = 'http://www.omdbapi.com/?apikey=33c97183'    
-
-    const title ='frozen'
-    const year  = 2019
-    const page  =''
-    const imdbID =''
+async function initializeDtailTextbox() {
+    console.log(imdb_id);
     
-    try {
-        const response = await fetch(`${API_URL}&t=${title}&y=${year}`)
-        if(!response.ok){
-            throw new Error('response error')
+    if (imdb_id) {
+        const detailstextbox = document.querySelector('details-textbox')
+        if(detailstextbox){
+            await detailstextbox.setImdbId(imdb_id)
         }
-        const data = await response.json()
-        rander(data)
-    } catch (error) {
-        new Error('error : '+ error)
-        console.error('문제가 발생했습니다',error)   
+    }else{
+        console.error('IMDb ID가 URL에 없습니다!');
     }
 }
-getApiData()
-
-function rander(data) {
-    const details = document.querySelector('.details')    
-    details.innerHTML = `
-    <div class="details__row">
-        <h4 class="text--title__detail">${data.Title}</h4>
-        <p class="text--contents__detail">${data.Year}</p>
-    </div>
-    `
-}
-
-// 클래스 선언
-// 디파인
+initializeDtailTextbox()
